implement add to cart

+++ lib/cart.action.ts
@@ -6,15 +6,22 @@ import { CartParams } from "@/types/types";
 import { NextResponse } from "next/server";
 import dbConnect from "./dbconnection";
 
-export async function addToCart({ categoryName, productId }: CartParams) {
+export async function addToCart({
+  categoryName,
+  price,
+  productId,
+}: CartParams) {
   const session = await auth();
   await dbConnect();
   try {
+    if (!session)
+      throw new Error("You need to authenticate to perform such actions");
     const [newCartItem] = await CartModel.create([
       {
         categoryName: categoryName,
         productId: productId,
         ownerId: session?.user?.id,
+        price,
       },
     ]);
 
@@ -43,7 +50,10 @@ export async function removeFromCart({
   productId,
 }: Pick<CartParams, "productId">) {
   await dbConnect();
+  const session = await auth();
   try {
+    if (!session)
+      throw new Error("You need to authenticate to perform such actions");
     const deletedCartItem = await CartModel.findOneAndDelete(
       { productId },
       { new: true },
@@ -63,9 +73,38 @@ export async function getCartItems() {
   const session = await auth();
   await dbConnect();
   try {
+    if (!session)
+      throw new Error("You need to authenticate to perform such actions");
     const carItems = await CartModel.find({
       ownerId: session?.user?.id,
-    }).select("productId ownerId");
+    });
+
+    if (!carItems) return { success: false, data: [] };
+
+    return { success: true, data: JSON.parse(JSON.stringify(carItems)) };
+  } catch (error) {
+    if (error instanceof Error)
+      return { success: false, message: error.message, status: 400 };
+    return { success: false, message: String(error), status: 400 };
+  }
+}
+
+export async function incrmentDecrementCart(productId: string) {
+  const session = await auth();
+  await dbConnect();
+  try {
+    if (!session)
+      throw new Error("You need to authenticate to perform such actions");
+    const carItems = await CartModel.findOneAndUpdate(
+      {
+        productId,
+      },
+      {
+        $inc: {
+          quantity: 1,
+        },
+      },
+    );
 
     if (!carItems) return { success: false, data: [] };
